requirements.txt
# Web Framework
Flask>=2.3.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# AI Providers
google-generativeai>=0.3.0
anthropic>=0.18.0
openai>=1.12.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# Database
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0
psycopg2-binary>=2.9.0

# Security - API key encryption
cryptography>=41.0.0

# Google Sheets Integration
gspread>=5.12.0
google-auth>=2.22.0

# Production Server
gunicorn>=21.0.0

# Utilities
pathlib2>=2.3.0
